MR : Feat: Add UserRepository.java | juyoung => repository

See merge request s09-webmobile3-sub2/S09P12A304!41

+++ BackEnd/ggong/src/main/java/com/a304/ggong/entity/User.java
@@ -15,7 +15,7 @@ import javax.persistence.*;
 public class User {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_no")
-    private int user_no;
+    private Long user_no;
 
     @Column(name = "nickname")
     private String name;
